git: Remove gpgSigner from worktree
